[backend] refactor, remove MentorIds from InviteExpertViewModel and WorkspaceDTO
