+++ HwProj.Common/HwProj.Models/AuthService/ViewModels/InviteExpertViewModel.cs
@@ -13,7 +13,5 @@ namespace HwProj.Models.AuthService.ViewModels
         public List<string> StudentIds { get; set; } = new List<string>();
 
         public List<long> HomeworkIds { get; set; } = new List<long>();
-
-        public List<string> MentorIds { get; set; } = new List<string>();
     }
 }
