+++ HwProj.CoursesService/HwProj.CoursesService.API/Models/Filter.cs
@@ -7,11 +7,11 @@ namespace HwProj.CoursesService.API.Models
     {
         [JsonProperty(PropertyName = "STUD")] 
         public List<string> StudentIds { get; set; }
-        
+
         [JsonProperty(PropertyName = "HMW")] 
         public List<long> HomeworkIds { get; set; }
         
         [JsonProperty(PropertyName = "MENT")]
         public List<string> MentorIds { get; set; }
     }
-}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
+}
